+++ src/components/hptracker/DropGroup.tsx
@@ -1,8 +1,10 @@
 import { Droppable } from "react-beautiful-dnd";
 import { DraggableTokenList } from "./TokenList.tsx";
 import OBR, { Item, Metadata } from "@owlbear-rodeo/sdk";
-import { SceneMetadata } from "../../helper/types.ts";
-import { sceneMetadata } from "../../helper/variables.ts";
+import { HpTrackerMetadata, SceneMetadata } from "../../helper/types.ts";
+import { characterMetadata, sceneMetadata } from "../../helper/variables.ts";
+import { updateHp } from "../../helper/hpHelpers.ts";
+import { updateAc } from "../../helper/acHelper.ts";
 
 type DropGroupProps = {
     title: string;
@@ -22,8 +24,103 @@ export const DropGroup = (props: DropGroupProps) => {
         }
         const ownMetadata: Metadata = {};
         ownMetadata[sceneMetadata] = hpTrackerSceneMetadata;
-        console.log(hpTrackerSceneMetadata);
         await OBR.scene.setMetadata(ownMetadata);
+    };
+
+    const getHpBar = () => {
+        const hpBars = props.list.map((token) => {
+            const metadata = token.metadata[characterMetadata] as HpTrackerMetadata;
+            return metadata.hpBar;
+        });
+        return hpBars.some((bar) => bar);
+    };
+
+    const toggleHpBar = async () => {
+        const current = getHpBar();
+        await OBR.scene.items.updateItems(props.list, (items) => {
+            items.forEach((item) => {
+                (item.metadata[characterMetadata] as HpTrackerMetadata).hpBar = !current;
+            });
+        });
+        for (const item of props.list) {
+            const data = item.metadata[characterMetadata] as HpTrackerMetadata;
+            await updateHp(item, { ...data, hpBar: !current });
+        }
+    };
+
+    const getHpOnMap = () => {
+        const hpMap = props.list.map((token) => {
+            const metadata = token.metadata[characterMetadata] as HpTrackerMetadata;
+            return metadata.hpOnMap;
+        });
+        return hpMap.some((map) => map);
+    };
+
+    const toggleHpOnMap = async () => {
+        const current = getHpOnMap();
+        await OBR.scene.items.updateItems(props.list, (items) => {
+            items.forEach((item) => {
+                (item.metadata[characterMetadata] as HpTrackerMetadata).hpOnMap = !current;
+            });
+        });
+        for (const item of props.list) {
+            const data = item.metadata[characterMetadata] as HpTrackerMetadata;
+            await updateHp(item, { ...data, hpOnMap: !current });
+        }
+    };
+
+    const getAcOnMap = () => {
+        const acMap = props.list.map((token) => {
+            const metadata = token.metadata[characterMetadata] as HpTrackerMetadata;
+            return metadata.acOnMap;
+        });
+        return acMap.some((map) => map);
+    };
+
+    const toggleAcOnMap = async () => {
+        const current = getAcOnMap();
+        await OBR.scene.items.updateItems(props.list, (items) => {
+            items.forEach((item) => {
+                (item.metadata[characterMetadata] as HpTrackerMetadata).acOnMap = !current;
+            });
+        });
+        for (const item of props.list) {
+            const data = item.metadata[characterMetadata] as HpTrackerMetadata;
+            await updateAc(item, { ...data, acOnMap: !current });
+        }
+    };
+    const getCanPlayersSee = () => {
+        const canSee = props.list.map((token) => {
+            const metadata = token.metadata[characterMetadata] as HpTrackerMetadata;
+            return metadata.canPlayersSee;
+        });
+        return canSee.some((see) => see);
+    };
+
+    const toggleCanPlayerSee = async () => {
+        const current = getCanPlayersSee();
+        await OBR.scene.items.updateItems(props.list, (items) => {
+            items.forEach((item) => {
+                (item.metadata[characterMetadata] as HpTrackerMetadata).canPlayersSee = !current;
+            });
+        });
+        for (const item of props.list) {
+            const data = item.metadata[characterMetadata] as HpTrackerMetadata;
+            await updateHp(item, { ...data, canPlayersSee: !current });
+            await updateAc(item, { ...data, canPlayersSee: !current });
+        }
+    };
+
+    const setInitiative = async () => {
+        await OBR.scene.items.updateItems(props.list, (items) => {
+            items.forEach((item) => {
+                const value =
+                    Math.floor(Math.random() * 20) +
+                    1 +
+                    (item.metadata[characterMetadata] as HpTrackerMetadata).stats.initiativeBonus;
+                (item.metadata[characterMetadata] as HpTrackerMetadata).initiative = value;
+            });
+        });
     };
 
     return (
@@ -33,7 +130,46 @@ export const DropGroup = (props: DropGroupProps) => {
             }`}
         >
             <div className={"group-title"}>
-                {props.title}{" "}
+                <div className={"group-name"}>
+                    <span>{props.title}</span>
+                </div>
+                <div className={"settings"}>
+                    <button
+                        title={"Toggle HP Bar visibility for GM and Players"}
+                        className={`toggle-button hp ${getHpBar() ? "on" : "off"}`}
+                        onClick={() => {
+                            toggleHpBar();
+                        }}
+                    />
+                    <button
+                        title={"Toggle HP displayed on Map"}
+                        className={`toggle-button map ${getHpOnMap() ? "on" : "off"}`}
+                        onClick={() => {
+                            toggleHpOnMap();
+                        }}
+                    />
+                    <button
+                        title={"Toggle AC displayed on Map"}
+                        className={`toggle-button ac ${getAcOnMap() ? "on" : "off"}`}
+                        onClick={async () => {
+                            toggleAcOnMap();
+                        }}
+                    />
+                    <button
+                        title={"Toggle HP/AC visibility for players"}
+                        className={`toggle-button players ${getCanPlayersSee() ? "on" : "off"}`}
+                        onClick={() => {
+                            toggleCanPlayerSee();
+                        }}
+                    />
+                </div>
+                <button
+                    title={"Roll Initiative (including initiative modifier from statblock)"}
+                    className={`toggle-button initiative-button`}
+                    onClick={() => {
+                        setInitiative();
+                    }}
+                />
                 <button
                     className={"hide-group"}
                     onClick={async () => {
@@ -41,19 +177,21 @@ export const DropGroup = (props: DropGroupProps) => {
                     }}
                 ></button>
             </div>
-            <div className={"drop-list"}>
-                <Droppable droppableId={props.title}>
-                    {(provided) => (
-                        <div ref={provided.innerRef} {...provided.droppableProps}>
-                            <DraggableTokenList
-                                tokens={props.list}
-                                selected={props.selected}
-                                metadata={props.metadata}
-                            />
-                            {provided.placeholder}
-                        </div>
-                    )}
-                </Droppable>
+            <div className={"drop-list-wrapper"}>
+                <div className={"drop-list"}>
+                    <Droppable droppableId={props.title}>
+                        {(provided) => (
+                            <div ref={provided.innerRef} {...provided.droppableProps}>
+                                <DraggableTokenList
+                                    tokens={props.list}
+                                    selected={props.selected}
+                                    metadata={props.metadata}
+                                />
+                                {provided.placeholder}
+                            </div>
+                        )}
+                    </Droppable>
+                </div>
             </div>
         </div>
     );